% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bmtcrr}
\alias{bmtcrr}
\title{Data on transplant patients}
\format{
A dataframe with 177 observations and 7 variables: \describe{
\item{Sex}{Gender of the individual} \item{D}{Disease: lymphoblastic or
myeloblastic leukemia, abbreviated as ALL and AML, respectively}
\item{Phase}{Phase at transplant (Relapse, CR1, CR2, CR3)} \item{Age}{Age
at the beginning of follow-up} \item{Status}{Status indicator: 0=censored,
1=relapse, 2=competing event} \item{Source}{Source of stem cells: bone
marrow and peripheral blood, coded as BM+PB, or peripheral blood only,
coded as PB} \item{ftime}{Failure time in months} }
}
\source{
Available at the following website:
\url{http://www.stat.unipg.it/luca/R/}
}
\usage{
bmtcrr
}
\description{
Data on patients who underwent haematopoietic stem cell transplantation for
acute leukemia.
}
\references{
Scrucca L, Santucci A, Aversa F. Competing risk analysis using R:
an easy guide for clinicians. Bone Marrow Transplant. 2007 Aug;40(4):381-7.
\doi{10.1038/sj.bmt.1705727}.
}
\keyword{datasets}
