% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{class}
\name{CompRisk-class}
\alias{CompRisk-class}
\alias{CompRisk}
\alias{summary}
\alias{summary,CompRisk-method}
\title{An S4 class to store the output of fitSmoothHazard}
\usage{
summary(object, ...)

\S4method{summary}{CompRisk}(object)
}
\arguments{
\item{object}{Object of class \code{CompRisk}}

\item{...}{Extra parameters}
}
\description{
This class inherits from \code{vglm-class}.
}
\section{Slots}{

\describe{
\item{\code{originalData}}{Data.frame containing the original data (i.e. before
case-base sampling). This is used by the \code{\link{absoluteRisk}}
function.}

\item{\code{typeEvents}}{Numeric factor which encodes the type of events being
considered (including censoring).}

\item{\code{timeVar}}{Character string giving the name of the time variable, as
appearing in \code{originalData}}

\item{\code{eventVar}}{Character string giving the name of the event variable, as
appearing in \code{originalData}}
}}

