% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eprchd}
\alias{eprchd}
\title{Estrogen plus Progestin and the Risk of Coronary Heart Disease (eprchd)}
\format{
A dataframe with 16608 observations and 3 variables:
\describe{ \item{time}{ Years (continuous) } \item{status}{
0=censored, 1=event } \item{treatment}{ placebo,
estPro}}
}
\usage{
eprchd
}
\description{
This data was reconstructed from the curves in figure 2
(Manson 2003).Compares placebo to hormone treatment.
}
\examples{
data("eprchd")
fit <- fitSmoothHazard(status ~ time + treatment, data = eprchd)
}
\references{
Manson, J. E., Hsia, J., Johnson, K. C., Rossouw, J. E., Assaf,
A. R., Lasser, N. L., ... & Strickland, O. L. (2003). Estrogen plus
progestin and the risk of coronary heart disease. New England Journal of
Medicine, 349(6), 523-534.
}
\keyword{datasets}
