% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitting.R
\name{fitSmoothHazard}
\alias{fitSmoothHazard}
\title{Fit smooth-in-time parametric hazard functions.}
\usage{
fitSmoothHazard(formula, data, link)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to
that class): a symbolic description of the model to be fitted. The details
of model specification are given under ‘Details’.}

\item{data}{an optional data frame, list or environment (or object coercible
by as.data.frame to a data frame) containing the variables in the model. If
not found in data, the variables are taken from environment(formula),
typically the environment from which glm is called.}

\item{link}{A character string, giving the specification for the model link
function.}
}
\value{
An object of class \code{caseBase}, which inherits from the classes
  \code{glm} and \code{lm}. As such, functions like \code{summary} and
  \code{coefficients} give familiar results.
}
\description{
Fit smooth-in-time parametric hazard functions.
}

