% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{masonTrialExtraction}
\alias{masonTrialExtraction}
\title{Data extracted from the mason trial}
\format{
A dataframe with 16608 observations and 3 variables:
}
\source{
taken from the following paper:
}
\usage{
masonTrialExtraction
}
\description{
Compares placebo to hormone treatment. THIS IS NOT EXTENSIVE. Will add
details later.
}
\references{
proper reference for the data.
}
\keyword{datasets}
