% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.popTimeExposure}
\alias{plot.popTimeExposure}
\title{Population Time Plot Stratified by Exposure Status}
\usage{
\method{plot}{popTimeExposure}(x, ..., ncol = 1, xlab = "Follow-up time",
  ylab = "Population", line.width = 1, line.colour = "grey80",
  point.size = 1, point.colour = "red", legend = FALSE,
  legend.position = c("bottom", "top", "left", "right"))
}
\arguments{
\item{x}{an object of class \code{popTimeExposure}. See
\code{\link{popTime}} for details.}

\item{...}{Ignored.}

\item{ncol}{Number of columns.}

\item{xlab, ylab, line.width, line.colour, point.size, point.colour, legend, legend.position}{See
\code{\link{par}}.}
}
\value{
a population time plot stratified by exposure status
}
\description{
\code{plot} method for objects of class \code{popTimeExposure}
}
\examples{
\dontrun{
DT <- read.csv(system.file("extdata", "bmtcrr.csv", package = "casebase"))
popTimeData <- popTime(data = DT, time = "ftime", exposure = "D")
# p is an object of class gg and ggplot
p <- plot(popTimeData)
# you can further modify the object using all ggplot2 functions
# here we modify the number of y-tick labels
p + scale_y_continuous(breaks = seq(0, max(popTimeData$data$ycoord), 10))
}
}
