% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampling.R
\name{sampleCaseBase}
\alias{sampleCaseBase}
\title{Create case-base dataset for use in fitting parametric hazard functions}
\usage{
sampleCaseBase(data, ratio = 10, type = c("uniform", "multinomial"))
}
\arguments{
\item{data}{The source dataset; see Details.}

\item{ratio}{Integer, giving the ratio of the size of the base series to that
of the case series. Defaults to 10.}

\item{type}{There are currently two sampling procedures available:
\code{uniform}, where person-moments are sampled uniformly across
individuals and follow-up time; and \code{multinomial}, where individuals
are sampled proportionally to their follow-up time.}
}
\value{
The function returns a dataset, with the same format as the source
  dataset, and where each row corresponds to a person-moment sampled from the
  case or the base series. otherwise)
}
\description{
This function implements the case-base sampling approach described in Hanley
and Miettinen, Int J Biostatistics 2009. It can be used to fit smooth-in-time
parametric functions easily, via logistic regression.
}
\details{
It is assumed that \code{data} contains two columns named \code{time} and
\code{event} (see the function \code{\link{Surv}} in the package
\code{survival}).
}

