% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitting.R
\name{fitSmoothHazard}
\alias{fitSmoothHazard}
\title{Fit smooth-in-time parametric hazard functions.}
\usage{
fitSmoothHazard(formula, data, time, event, link = "logit", ...)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to
that class): a symbolic description of the model to be fitted. The details
of model specification are given under Details.}

\item{data}{a data frame, list or environment containing the variables in the
model. If not found in data, the variables are taken from
\code{environment(formula)}, typically the environment from which
\code{fitSmoothHazard} is called.}

\item{time}{a character string giving the name of the time variable. See
Details.}

\item{event}{a character string giving the name of the event variable. See
Details.}

\item{link}{A character string, which gives the specification for the model
link function. Default is the \code{logit} link.}

\item{...}{Additional parameters passed to \code{\link{sampleCaseBase}}. If
\code{data} inherits from the class \code{cbData}, then these parameters
are ignored.}
}
\value{
An object of class \code{caseBase}, which inherits from the classes
  \code{glm} and \code{lm}. As such, functions like \code{summary} and
  \code{coefficients} give familiar results.
}
\description{
Miettinen and Hanley (2009) explained how case-base sampling can be used to
estimate smooth-in-time parametric hazard functions. The idea is to sample
person-moments, which may or may not correspond to an event, and then fit the
hazard using logistic regression.
}
\details{
The object \code{data} should either be the output of the function
\link{\code{sampleCaseBase}} or the source dataset on which case-base
sampling will be performed. In the latter case, it is assumed that
\code{data} contains the two columns corresponding to the supplied time and
event variables. If either the \code{time} or \code{event} argument is
missing, the function looks for columns named \code{"time"}, \code{"event"},
or \code{"status"}.
}

