% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.popTime}
\alias{plot.popTime}
\title{Population Time Plot}
\usage{
plot.popTime(object, xlab = "Follow-up time", ylab = "Population",
  line.width = 1, line.colour = "grey80", point.size = 1,
  point.colour = "red", legend = FALSE, legend.position = c("bottom",
  "top", "left", "right"))
}
\arguments{
\item{object}{an object of class \code{popTime}. See \code{\link{popTime}}
for details}
}
\value{
a population time plot
}
\description{
\code{plot} method for objects of class \code{popTime}
}

