% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.popTimeExposure}
\alias{plot.popTimeExposure}
\title{Population Time Plot Stratified by Exposure Status}
\usage{
\method{plot}{popTimeExposure}(object, ncol = 1, xlab = "Follow-up time",
  ylab = "Population", line.width = 1, line.colour = "grey80",
  point.size = 1, point.colour = "red", legend = FALSE,
  legend.position = c("bottom", "top", "left", "right"))
}
\arguments{
\item{object}{an object of class \code{popTimeExposure}. See
\code{\link{popTime}} for details}
}
\value{
a population time plot stratified by exposure status
}
\description{
\code{plot} method for objects of class \code{popTimeExposure}
}

