% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sampleCaseBase}
\alias{sampleCaseBase}
\title{Create case-base dataset for use in fitting parametric hazard functions}
\usage{
sampleCaseBase(data, time, event, ratio = 10, type = c("uniform",
  "multinomial"), cmprisk = FALSE)
}
\arguments{
\item{data}{a data.frame or data.table containing the source dataset.}

\item{time}{a character string giving the name of the time variable. See
Details.}

\item{event}{a character string giving the name of the event variable. See
Details.}

\item{ratio}{Integer, giving the ratio of the size of the base series to that
of the case series. Defaults to 10.}

\item{type}{There are currently two sampling procedures available:
\code{uniform}, where person-moments are sampled uniformly across
individuals and follow-up time; and \code{multinomial}, where individuals
are sampled proportionally to their follow-up time.}

\item{cmprisk}{Logical. Indicates whether we have multiple event types and
that we want to consider some of them as competing risks.}
}
\value{
The function returns a dataset, with the same format as the source
  dataset, and where each row corresponds to a person-moment sampled from the
  case or the base series. otherwise)
}
\description{
This function implements the case-base sampling approach described in Hanley
and Miettinen, Int J Biostatistics 2009. It can be used to fit smooth-in-time
parametric functions easily, via logistic regression.
}
\details{
It is assumed that \code{data} contains the two columns corresponding to the
supplied time and event variables. If either the \code{time} or \code{event}
argument is missing, the function looks for columns named \code{"time"},
\code{"event"}, or \code{"status"}.
}

