% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{support}
\alias{support}
\title{Study to Understand Prognoses Preferences Outcomes and Risks of Treatment (SUPPORT)}
\format{A dataframe with originally 1000 observations, but after imputation resulted in 474 observations and 35 covariates: \describe{ \item{Age}{ Stores a double representing age. } \item{death}{ Death at any time up to NDI date: 31DEC94. } \item{sex}{ 0=female, 1=male. } \item{hospdead}{ 1=dead in hospital, 0= has not died in hospital. } \item{slos}{ Days from study entry to discharge. } \item{d.time}{ days of follow-up. } \item{dzgroup}{ Each level of dzgroup: ARF/MOSF w/Sepsis, COPD, CHF, Cirrhosis, Coma, Colon Cancer, Lung Cancer, MOSF with malignancy. } \item{dzclass}{ ARF/MOSF, COPD/CHF/Cirrhosis, Coma and cancer disease classes. } \item{num.co}{ the number of comorbidities. } \item{edu}{ years of education of patient. } \item{income}{ bins individuals based on income bracket. Under $11k, $11-$25k, $25k-$50k or more than $50k. } \item{scoma}{ The SUPPORT coma score based on Glasgow D3. } \item{charges}{ How much the hospital charged for their services. } \item{totcst}{ Total RCC cost. } \item{totmcst}{ Total micro-cost. } \item{avtisst}{ Average TISS, days 3-25. } \item{race}{ Indicates race. White, black, asian, hispanic or other. } \item{meanbp}{ Mean Arterial Blood Pressure Day 3. } \item{wblc}{ White blood cell count on day 3. } \item{hrt}{ Heart rate day 3. } \item{resp}{ Respiration Rate day 3. } \item{temp}{ Temperature, in celcius, on day 3. } \item{pafi}{ PaO2/(0.01*FiO2) Day 3. } \item{alb}{ Serum albumin day 3. } \item{bili}{ Bilirubin Day 3. } \item{crea}{ Serum creatinine day 3. } \item{sod}{ Serum sodium day 3. } \item{ph}{ Serum pH (in arteries) day 3. } \item{glucose}{ Serum glucose day 3. } \item{bun}{  BUN day 3. } \item{urine}{ urine output day 3. } \item{adlp}{ ADL patient day 3. } \item{adls}{ ADL surrogate day 3. } \item{sfdm2}{ Ordinal variable that describes a set of mutually exclusive measures, with increasing severity. no(M2 and SIP pres) marks that the patient was able to survive for 2 months before an interview, go through a 2-month interview and have no signs of moderate to severe functional disability. adl>=4 (>=5 if sur) marks that the patient was unable to do 4 or more activities of daily living, or 5 or more activities of daily living if a surrogate was present at the meeting at month 2 after study entry.  SIP>=30 marks that the patient’s Sickness Impact Profile score was greater than 30. Coma or Intub marks that the patient is in a coma or intubated. <2 mo. follow-up marks that the patient did not make it through stage two of follow up.  } \item{adlsc}{ Imputed ADL calibrated to surrogate, if a surrogate was used for a follow up.  NEED TO DESCRIBE A FEW ACRONYMS}
}}
\source{
Available at the following website: \url{http://biostat.mc.vanderbilt.edu/wiki/pub/Main/DataSets/support.tsv}
}
\usage{
support
}
\description{
The support datasets track four response variables: hospital death, severe functional disability, hospital costs and time until death, where the patients are followed for up to 5.56 years.
}
\details{
Data from the first phase of the SUPPORT study. We imputed the data-set using predictive mean matching, which resulted in 474 complete observations.
}
\examples{
\dontrun{
library(splines)
library(casebase)
data(support)
mod_cb <- casebase::fitSmoothHazard(death ~ bs(d.time)+.-d.time, df = 3 +
                                   time = "d.time", event="death",
                                   data = support,
                                   ratio = 100)
}
}
\references{
Knaus WA, Harrell FE, Lynn J et al. (1995): The SUPPORT prognostic model: Objective estimates of survival for seriously ill hospitalized adults. Annals of Internal Medicine 122:191-203.

http://biostat.mc.vanderbilt.edu/wiki/Main/SupportDesc

http://biostat.mc.vanderbilt.edu/wiki/pub/Main/DataSets/Csupport.html
}
\keyword{datasets}
